refactor(templates): rename Commerce shell example to "Top Nav"
Rename the slug shell-commerce → shell-top-nav, set the display name to
"Top Nav", and update the docsite gallery component map.
