+++ apps/docsite/src/components/templateComponents.ts
@@ -117,9 +117,8 @@ export const TEMPLATE_COMPONENTS: Record<
   'side-gallery': lazy(
     () => import('../../../../packages/cli/templates/pages/side-gallery/page'),
   ),
-  'shell-commerce': lazy(
-    () =>
-      import('../../../../packages/cli/templates/pages/shell-commerce/page'),
+  'shell-top-nav': lazy(
+    () => import('../../../../packages/cli/templates/pages/shell-top-nav/page'),
   ),
   table: lazy(
     () => import('../../../../packages/cli/templates/pages/table/page'),
