--- packages/cli/templates/pages/shell-commerce/template.doc.mjs
@@ -1,12 +1,0 @@
-// Copyright (c) Meta Platforms, Inc. and affiliates.
-
-/** @type {import('../../../../core/src/docs-types').TemplateDoc} */
-export const doc = {
-  type: 'page',
-  name: 'Commerce Marketing Shell',
-  displayName: 'Commerce Marketing Shell',
-  description:
-    'A storefront shell with a centered TopNav — Shop and Brands mega menus (a 2x4 item grid plus a featured card each), Sale/Service links, and a Checkout button — over static grey-card hero and category sections.',
-  isReady: true,
-  category: 'Shell - Top Nav',
-};
